+++ src/identifier_ref.rs
@@ -34,7 +34,6 @@ use crate::{ClassTag, Error, Id, PCTag};
 use num::{FromPrimitive, PrimInt};
 use std::borrow::Borrow;
 use std::cmp::Ordering;
-use std::convert::TryFrom;
 use std::mem;
 
 pub const LONG_FLAG: u8 = 0x1f;
@@ -54,49 +53,6 @@ pub const PC_MASK: u8 = 0x20;
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct IdRef {
     bytes: [u8],
-}
-
-impl<'a> TryFrom<&'a [u8]> for &'a IdRef {
-    type Error = Error;
-
-    /// Parses `bytes` starting with identifier, and tries to provide a reference to `IdRef`.
-    ///
-    /// This function ignores the extra octet(s) at the end if any.
-    ///
-    /// This function is the same as [`IdRef::parse`].
-    ///
-    /// # Warnings
-    ///
-    /// ASN.1 reserves some universal identifier numbers and they should not be used, however,
-    /// this function ignores that. For example, number 15 (0x0f) is reserved for now, but this
-    /// functions returns `Ok`.
-    ///
-    /// [Read more](std::convert::TryFrom::try_from)
-    fn try_from(bytes: &'a [u8]) -> Result<Self, Self::Error> {
-        IdRef::parse(bytes)
-    }
-}
-
-impl<'a> TryFrom<&'a mut [u8]> for &'a mut IdRef {
-    type Error = Error;
-
-    /// Parses `bytes` starting with identifier, and tries to provide a mutable reference to
-    /// `IdRef`.
-    ///
-    /// This function ignores the extra octet(s) at the end if any.
-    ///
-    /// This function is the same as [`IdRef::parse_mut`].
-    ///
-    /// # Warnings
-    ///
-    /// ASN.1 reserves some universal identifier numbers and they should not be used, however,
-    /// this function ignores that. For example, number 15 (0x0f) is reserved for now, but this
-    /// functions returns `Ok`.
-    ///
-    /// [Read more](std::convert::TryFrom::try_from)
-    fn try_from(bytes: &'a mut [u8]) -> Result<Self, Self::Error> {
-        IdRef::parse_mut(bytes)
-    }
 }
 
 /// Ommits the extra octets at the end of `bytes` and returns octets just one 'ASN.1 Identifier.'
@@ -1454,12 +1410,12 @@ mod tests {
                 {
                     let first = cl as u8 + pc as u8 + 0;
                     let bytes: &[u8] = &[first];
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
 
                     let first = cl as u8 + pc as u8 + 0x1e;
                     let bytes: &[u8] = &[first];
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
                 }
 
@@ -1468,11 +1424,11 @@ mod tests {
                 // 2 bytes
                 {
                     let bytes: &[u8] = &[first, 0x1f];
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
 
                     let bytes: &[u8] = &[first, 0x7f];
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
                 }
 
@@ -1486,7 +1442,7 @@ mod tests {
                     bytes[len - 1] = 0x00;
 
                     let bytes: &[u8] = &bytes;
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
 
                     let mut bytes = vec![first];
@@ -1496,7 +1452,7 @@ mod tests {
                     bytes[len - 1] = 0x7f;
 
                     let bytes: &[u8] = &bytes;
-                    let id = <&IdRef>::try_from(bytes).unwrap();
+                    let id = IdRef::parse(bytes).unwrap();
                     assert_eq!(bytes, id.as_bytes());
                 }
             }
@@ -1512,11 +1468,11 @@ mod tests {
                 // 2 bytes
                 {
                     let bytes: &[u8] = &[first, 0x00];
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::RedundantBytes, e);
 
                     let bytes: &[u8] = &[first, 0x1e];
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::RedundantBytes, e);
                 }
 
@@ -1529,7 +1485,7 @@ mod tests {
                     bytes[len - 1] = 0x00;
 
                     let bytes: &[u8] = &bytes;
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::RedundantBytes, e);
 
                     let mut bytes = vec![first];
@@ -1540,7 +1496,7 @@ mod tests {
                     bytes[len - 1] = 0x7f;
 
                     let bytes: &[u8] = &bytes;
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::RedundantBytes, e);
                 }
             }
@@ -1552,7 +1508,7 @@ mod tests {
         // Empty
         {
             let bytes: &[u8] = &[];
-            let e = <&IdRef>::try_from(bytes).unwrap_err();
+            let e = IdRef::parse(bytes).unwrap_err();
             assert_eq!(Error::UnTerminatedBytes, e);
         }
 
@@ -1563,7 +1519,7 @@ mod tests {
                 // 1 bytes
                 {
                     let bytes: &[u8] = &[first];
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::UnTerminatedBytes, e);
                 }
 
@@ -1576,7 +1532,7 @@ mod tests {
                     bytes[1] = 0x81;
 
                     let bytes: &[u8] = &bytes;
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::UnTerminatedBytes, e);
 
                     let mut bytes = vec![first];
@@ -1585,7 +1541,7 @@ mod tests {
                     }
 
                     let bytes: &[u8] = &bytes;
-                    let e = <&IdRef>::try_from(bytes).unwrap_err();
+                    let e = IdRef::parse(bytes).unwrap_err();
                     assert_eq!(Error::UnTerminatedBytes, e);
                 }
             }